0.7.0 starting point (#38)
0.7.0 starting point 

* Created a bare bones version of wakepy for 0.7.0 update
* Removed all new "dev branch additions"
* This provides a starting point for the new API for wakepy version 0.7.0
* Deprecated keepwake, set_keepawake and unset_keepawake
