+++ wakepy/__init__.py
@@ -1,4 +1,4 @@
-from .keepawake import keepawake, set_keepawake, unset_keepawake
+from ._deprecated import keepawake, set_keepawake, unset_keepawake
 
 __version__ = "0.7.0dev"
