+++ wakepy/_deprecated/_linux.py
@@ -7,44 +7,13 @@ Requires jeepney. Install with:
 See also:
     https://people.freedesktop.org/~hadess/idle-inhibition-spec/re01.html
 
-
-The inhibitors can be checked with
-    dbus-send --print-reply --dest=org.gnome.SessionManager /org/gnome/SessionManager org.gnome.SessionManager.GetInhibitors
-
-For example:
-
-$ dbus-send --print-reply --dest=org.gnome.SessionManager /org/gnome/SessionManager org.gnome.SessionManager.GetInhibitors
-method return time=1681201482.075825 sender=:1.25 -> destination=:1.559 serial=465 reply_serial=2
-   array [
-      object path "/org/gnome/SessionManager/Inhibitor7"
-   ]
-
-This can be examined further with
-
-dbus-send --print-reply --dest=org.gnome.SessionManager /org/gnome/SessionManager/InhibitorN org.gnome.SessionManager.Inhibitor.GetAppId
-
-For example:
-
-niko@niko-ubuntu-home:~$ dbus-send --print-reply --dest=org.gnome.SessionManager /org/gnome/SessionManager/Inhibitor7 org.gnome.SessionManager.Inhibitor.GetAppId
-method return time=1681203094.197509 sender=:1.25 -> destination=:1.684 serial=480 reply_serial=2
-   string "wakepy"
 """
-
-from ...exceptions import KeepAwakeError
-from ...inhibitor import KeepAwakeInfo, InhibitorInfo
 
 # There will be imported from jeepney when needed
 new_method_call = None
 DBusAddress = None
 open_dbus_connection = None
 
-# Values for wakepy.core (for error handling / logging)
-PRINT_NAME = "jeepney (dbus)"
-REQUIREMENTS = [
-    "session message bus (dbus-daemon) running",
-    "DBUS_SESSION_BUS_ADDRESS set",
-    "jeepney (python package)",
-]
 
 # The keepawake setter / unsetter object
 # Will be None until first call of methods
@@ -58,7 +27,7 @@ def import_jeepney():
         from jeepney import new_method_call, DBusAddress
         from jeepney.io.blocking import open_dbus_connection
     except Exception as e:
-        raise KeepAwakeError("Could not import jeepney!") from e
+        raise NotImplementedError("Could not import jeepney!") from e
 
 
 def get_connection(bus="SESSION"):
@@ -69,12 +38,12 @@ def get_connection(bus="SESSION"):
         return open_dbus_connection(bus=bus)
     except Exception as e:
         if "DBUS_SESSION_BUS_ADDRESS" in str(e):
-            raise KeepAwakeError(
+            raise NotImplementedError(
                 "DBUS_SESSION_BUS_ADDRESS environment variable not set! "
                 "If running in subprocess, make sure to pass the DBUS_SESSION_BUS_ADDRESS "
                 "environment variable."
             ) from e
-        raise KeepAwakeError(
+        raise NotImplementedError(
             f"Could not set dbus connection to {bus} message bus.\n"
             f"{e.__class__.__name__}: {str(e)}"
         ) from e
@@ -121,7 +90,7 @@ class KeepAwakeSetter:
 
     def unset_keepawake(self):
         if self.inhibit_cookie is None:
-            raise KeepAwakeError("You must set_keepawake before unsetting!")
+            raise NotImplementedError("You must set_keepawake before unsetting!")
 
         msg_uninhibit = new_method_call(
             self.screensaver,
@@ -149,59 +118,3 @@ def set_keepawake(keep_screen_awake=False):
 def unset_keepawake():
     setter = get_setter()
     setter.unset_keepawake()
-
-
-def _get_inhibitor_and_reason(connection, obj_path: str):
-    addr = DBusAddress(
-        obj_path,
-        bus_name="org.gnome.SessionManager",
-        interface="org.gnome.SessionManager.Inhibitor",
-    )
-
-    def _get(method):
-        msg = new_method_call(addr, method, "", tuple())
-        reply = connection.send_and_get_reply(msg)
-        return reply.body[0]
-
-    name = _get("GetAppId")
-    reason = _get("GetReason")
-
-    return name, reason
-
-
-def check_keepawake() -> KeepAwakeInfo:
-    """Checks keepawake status (dbus). Experimental."""
-    try:
-        return check_keepawake_gnome()
-    except Exception as e:
-        raise KeepAwakeError(f"Cannot check keepawake! Reason: {str(e)}") from e
-
-
-def check_keepawake_gnome() -> KeepAwakeInfo:
-    """Checks keepawake status on gnome (dbus). Experimental."""
-    info = KeepAwakeInfo()
-
-    if DBusAddress is None:
-        import_jeepney()
-
-    gnome_session_manager_addr = DBusAddress(
-        "/org/gnome/SessionManager",  # DBus object path
-        bus_name="org.gnome.SessionManager",
-        interface="org.gnome.SessionManager",
-    )
-    msg = new_method_call(
-        gnome_session_manager_addr,  # DBusAddress
-        "GetInhibitors",  # Method
-        "",  # No input for method
-        tuple(),
-    )
-    connection = get_connection()
-    reply = connection.send_and_get_reply(msg)
-
-    for inhibitor_tuple in reply.body[0]:
-        inhibitor_name, inhibitor_reason = _get_inhibitor_and_reason(
-            connection, inhibitor_tuple
-        )
-        info.add(InhibitorInfo(inhibitor_name, inhibitor_reason))
-
-    return info
