--- wakepy/exceptions.py
@@ -1,2 +1,0 @@
-class KeepAwakeError(Exception):
-    ...
